requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.2.0
langchain-community>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
# redis[asyncio]>=5.0.0  # replaced by Google Sheets cache
# crawl4ai>=0.3.0  # replaced by httpx
# playwright>=1.44.0  # replaced by httpx
# playwright-stealth>=1.0.0  # replaced by httpx
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.7.0
beautifulsoup4>=4.12.0
ddgs>=9.0.0
pyairtable>=2.0.0
hubspot-api-client>=8.0.0
colorama>=0.4.6
unstructured>=0.10.0
chromadb>=0.4.0
langchain-chroma>=0.1.0
